cloud/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo[srv]==4.10.1
bcrypt==4.1.2
stripe==7.9.0
pyjwt==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
certifi 